+++ packages/backend/src/fetch-data/index.ts
@@ -1,6 +1,6 @@
 export * from './fetch-food-data';
 
-export * from './fetch-recipes-data';
+export * from './admin-fetch-recipes-data';
 
 export * from './fetch-season-data';
 
@@ -8,4 +8,4 @@ export * from './fetch-season-with-food-data';
 
 export * from './fetch-season-with-recipe-data';
 
-export * from './fetch-user-data';
+export * from './admin-fetch-user-data';
